Code from MSc thesis, typechecking with current Agda.
For reference, I obtained this branch using the "easy mode" for
"squashing" commits:

http://makandracards.com/makandra/527-squash-several-git-commits-into-a-single-commit

However, I did not do it in master but in this branch, and made a
single commit after resetting to the first and staging what I wanted
by hand.

The "hard mode" required manual intervention, and I was worried it
would also rewrite master's history.