Recursive search of subtyping witnesses!
Thanks to Guillaume's enhancement, we now have (local) instance search
for subtyping witnesses.

As an experiment, I used this in two ways.

In IntroMod, I mimicked Data Types à la Carte's :<:. Here the relation
is symmetric, though.

In Sub and its users (Eff, MVCExample), I tried to keep my previous
workarounds: first I generate a list of pointed types (a Record)
corresponding to all the subtyping witnesses in Agda (smartSubs), then
I open them in the current scope using AD.Instances.

It seems to work as before, but now we don't have any static size
limits on the Record. I do not know how it would compare in time and
space with an ad-hoc module that provides all the recursive instances,
in particular in terms of laziness and garbage
collection-friendliness.

I should probably try.